+++ client/Packages/com.beamable/CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Content deletion popup opens as separate window
 - An optional `forceRefresh` parameter to all subscribable SDK's `GetCurrent()` method that forces a network request
-
+- `API.Instance.Requester` is now an `IBeamableRequester`
 
 ## [0.16.1]
 ### Fixed